dedupe some registrar controller references
