--- apps/web/src/contracts/Usernames/constants.ts
@@ -1,442 +1,0 @@
-export const baseRegistrarSmartContractAddress =
-  (process.env.NEXT_PUBLIC_BASE_REGISTRAR_SMART_CONTRACT_ADDRESS as `0x${string}`) || '0x';
-
-export const baseRegistrarContractABI = [
-  {
-    inputs: [
-      { internalType: 'contract BaseRegistrar', name: 'base_', type: 'address' },
-      { internalType: 'contract IPriceOracle', name: 'prices_', type: 'address' },
-      { internalType: 'contract IReverseRegistrar', name: 'reverseRegistrar_', type: 'address' },
-      { internalType: 'address', name: 'owner_', type: 'address' },
-      { internalType: 'bytes32', name: 'rootNode_', type: 'bytes32' },
-      { internalType: 'string', name: 'rootName_', type: 'string' },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'target', type: 'address' }],
-    name: 'AddressEmptyCode',
-    type: 'error',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'account', type: 'address' }],
-    name: 'AddressInsufficientBalance',
-    type: 'error',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'sender', type: 'address' }],
-    name: 'AlreadyClaimedWithDiscount',
-    type: 'error',
-  },
-  { inputs: [], name: 'AlreadyInitialized', type: 'error' },
-  {
-    inputs: [{ internalType: 'uint256', name: 'duration', type: 'uint256' }],
-    name: 'DurationTooShort',
-    type: 'error',
-  },
-  { inputs: [], name: 'FailedInnerCall', type: 'error' },
-  {
-    inputs: [{ internalType: 'bytes32', name: 'key', type: 'bytes32' }],
-    name: 'InactiveDiscount',
-    type: 'error',
-  },
-  { inputs: [], name: 'InsufficientValue', type: 'error' },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'key', type: 'bytes32' },
-      { internalType: 'bytes', name: 'data', type: 'bytes' },
-    ],
-    name: 'InvalidDiscount',
-    type: 'error',
-  },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'key', type: 'bytes32' },
-      { internalType: 'uint256', name: 'amount', type: 'uint256' },
-    ],
-    name: 'InvalidDiscountAmount',
-    type: 'error',
-  },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'key', type: 'bytes32' },
-      { internalType: 'address', name: 'validator', type: 'address' },
-    ],
-    name: 'InvalidValidator',
-    type: 'error',
-  },
-  {
-    inputs: [{ internalType: 'string', name: 'name', type: 'string' }],
-    name: 'NameNotAvailable',
-    type: 'error',
-  },
-  { inputs: [], name: 'NewOwnerIsZeroAddress', type: 'error' },
-  { inputs: [], name: 'NoHandoverRequest', type: 'error' },
-  { inputs: [], name: 'ResolverRequiredWhenDataSupplied', type: 'error' },
-  {
-    inputs: [{ internalType: 'address', name: 'token', type: 'address' }],
-    name: 'SafeERC20FailedOperation',
-    type: 'error',
-  },
-  { inputs: [], name: 'TransferFailed', type: 'error' },
-  {
-    inputs: [{ internalType: 'bytes32', name: 'node', type: 'bytes32' }],
-    name: 'Unauthorised',
-    type: 'error',
-  },
-  { inputs: [], name: 'Unauthorized', type: 'error' },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'bytes32', name: 'discountKey', type: 'bytes32' },
-      {
-        components: [
-          { internalType: 'bool', name: 'active', type: 'bool' },
-          { internalType: 'address', name: 'discountValidator', type: 'address' },
-          { internalType: 'uint256', name: 'discount', type: 'uint256' },
-        ],
-        indexed: false,
-        internalType: 'struct RegistrarController.DiscountDetails',
-        name: 'details',
-        type: 'tuple',
-      },
-    ],
-    name: 'DiscountUpdated',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: 'payee', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: 'price', type: 'uint256' },
-    ],
-    name: 'ETHPaymentProcessed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'string', name: 'name', type: 'string' },
-      { indexed: true, internalType: 'bytes32', name: 'label', type: 'bytes32' },
-      { indexed: true, internalType: 'address', name: 'owner', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: 'expires', type: 'uint256' },
-    ],
-    name: 'NameRegistered',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'string', name: 'name', type: 'string' },
-      { indexed: true, internalType: 'bytes32', name: 'label', type: 'bytes32' },
-      { indexed: false, internalType: 'uint256', name: 'expires', type: 'uint256' },
-    ],
-    name: 'NameRenewed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverCanceled',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverRequested',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: 'oldOwner', type: 'address' },
-      { indexed: true, internalType: 'address', name: 'newOwner', type: 'address' },
-    ],
-    name: 'OwnershipTransferred',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: 'registrant', type: 'address' },
-      { indexed: true, internalType: 'bytes32', name: 'discountKey', type: 'bytes32' },
-    ],
-    name: 'RegisteredWithDiscount',
-    type: 'event',
-  },
-  {
-    inputs: [],
-    name: 'MIN_NAME_LENGTH',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'MIN_REGISTRATION_DURATION',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'string', name: 'name', type: 'string' }],
-    name: 'available',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'cancelOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'completeOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'string', name: 'name', type: 'string' },
-      { internalType: 'uint256', name: 'duration', type: 'uint256' },
-      { internalType: 'bytes32', name: 'discountKey', type: 'bytes32' },
-    ],
-    name: 'discountRentPrice',
-    outputs: [{ internalType: 'uint256', name: 'price', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'string', name: 'name', type: 'string' },
-          { internalType: 'address', name: 'owner', type: 'address' },
-          { internalType: 'uint256', name: 'duration', type: 'uint256' },
-          { internalType: 'address', name: 'resolver', type: 'address' },
-          { internalType: 'bytes[]', name: 'data', type: 'bytes[]' },
-          { internalType: 'bool', name: 'reverseRecord', type: 'bool' },
-        ],
-        internalType: 'struct RegistrarController.RegisterRequest',
-        name: 'request',
-        type: 'tuple',
-      },
-      { internalType: 'bytes32', name: 'discountKey', type: 'bytes32' },
-      { internalType: 'bytes', name: 'validationData', type: 'bytes' },
-    ],
-    name: 'discountedRegister',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '', type: 'address' }],
-    name: 'discountedRegistrants',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
-    name: 'discounts',
-    outputs: [
-      { internalType: 'bool', name: 'active', type: 'bool' },
-      { internalType: 'address', name: 'discountValidator', type: 'address' },
-      { internalType: 'uint256', name: 'discount', type: 'uint256' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'getActiveDiscounts',
-    outputs: [
-      {
-        components: [
-          { internalType: 'bool', name: 'active', type: 'bool' },
-          { internalType: 'address', name: 'discountValidator', type: 'address' },
-          { internalType: 'uint256', name: 'discount', type: 'uint256' },
-        ],
-        internalType: 'struct RegistrarController.DiscountDetails[]',
-        name: '',
-        type: 'tuple[]',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address[]', name: 'addresses', type: 'address[]' }],
-    name: 'hasRegisteredWithDiscount',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'owner',
-    outputs: [{ internalType: 'address', name: 'result', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'ownershipHandoverExpiresAt',
-    outputs: [{ internalType: 'uint256', name: 'result', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'prices',
-    outputs: [{ internalType: 'contract IPriceOracle', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '_token', type: 'address' },
-      { internalType: 'address', name: '_to', type: 'address' },
-      { internalType: 'uint256', name: '_amount', type: 'uint256' },
-    ],
-    name: 'recoverFunds',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'string', name: 'name', type: 'string' },
-          { internalType: 'address', name: 'owner', type: 'address' },
-          { internalType: 'uint256', name: 'duration', type: 'uint256' },
-          { internalType: 'address', name: 'resolver', type: 'address' },
-          { internalType: 'bytes[]', name: 'data', type: 'bytes[]' },
-          { internalType: 'bool', name: 'reverseRecord', type: 'bool' },
-        ],
-        internalType: 'struct RegistrarController.RegisterRequest',
-        name: 'request',
-        type: 'tuple',
-      },
-    ],
-    name: 'register',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'string', name: 'name', type: 'string' },
-      { internalType: 'uint256', name: 'duration', type: 'uint256' },
-    ],
-    name: 'registerPrice',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'string', name: 'name', type: 'string' },
-      { internalType: 'uint256', name: 'duration', type: 'uint256' },
-    ],
-    name: 'renew',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'renounceOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'string', name: 'name', type: 'string' },
-      { internalType: 'uint256', name: 'duration', type: 'uint256' },
-    ],
-    name: 'rentPrice',
-    outputs: [
-      {
-        components: [
-          { internalType: 'uint256', name: 'base', type: 'uint256' },
-          { internalType: 'uint256', name: 'premium', type: 'uint256' },
-        ],
-        internalType: 'struct IPriceOracle.Price',
-        name: 'price',
-        type: 'tuple',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'requestOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'reverseRegistrar',
-    outputs: [{ internalType: 'contract IReverseRegistrar', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'rootName',
-    outputs: [{ internalType: 'string', name: '', type: 'string' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'rootNode',
-    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'key', type: 'bytes32' },
-      {
-        components: [
-          { internalType: 'bool', name: 'active', type: 'bool' },
-          { internalType: 'address', name: 'discountValidator', type: 'address' },
-          { internalType: 'uint256', name: 'discount', type: 'uint256' },
-        ],
-        internalType: 'struct RegistrarController.DiscountDetails',
-        name: 'details',
-        type: 'tuple',
-      },
-    ],
-    name: 'setDiscountDetails',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'string', name: 'name', type: 'string' }],
-    name: 'valid',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'pure',
-    type: 'function',
-  },
-  { inputs: [], name: 'withdrawETH', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-];
